Release 0.8

+++ ptypy/accelerate/ocl_pyopencl/ocl_fft.py
@@ -174,7 +174,7 @@ class FFT_2D_ocl_reikna(object):
 
         # attach scaling
         from reikna.transformations import mul_param
-        sc = mul_param(array, np.float)
+        sc = mul_param(array, float)
         ftreikna.parameter.output.connect(sc, sc.input, out=sc.output, scale=sc.param)
         iscale = np.sqrt(np.prod(array.shape[-2:])) if symmetric else 1.0
         scale = 1.0 / iscale
